fix: Fix object assertion error
