+++ internal/ncloud/convert.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 	"unicode"
 
+	"github.com/NaverCloudPlatform/terraform-plugin-codegen-framework/internal/util"
 	"github.com/NaverCloudPlatform/terraform-plugin-codegen-spec/resource"
 )
 
@@ -15,17 +16,16 @@ func Gen_ConvertOAStoTFTypes(data resource.Attributes) (string, string, error) {
 
 	for _, val := range data {
 		n := val.Name
-		fmt.Println(n)
 
 		if val.String != nil {
-			s = s + fmt.Sprintf(`dto.%[1]s = types.StringValue(data["%[2]s"].(string))`, CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
-			m = m + fmt.Sprintf("%[1]s         types.String `json:\"%[2]s\"`", CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			s = s + fmt.Sprintf(`dto.%[1]s = types.StringValue(data["%[2]s"].(string))`, util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			m = m + fmt.Sprintf("%[1]s         types.String `json:\"%[2]s\"`", util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
 		} else if val.Bool != nil {
-			s = s + fmt.Sprintf(`dto.%[1]s = types.BoolValue(data["%[2]s"].(bool))`, CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
-			m = m + fmt.Sprintf("%[1]s         types.Bool `json:\"%[2]s\"`", CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			s = s + fmt.Sprintf(`dto.%[1]s = types.BoolValue(data["%[2]s"].(bool))`, util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			m = m + fmt.Sprintf("%[1]s         types.Bool `json:\"%[2]s\"`", util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
 		} else if val.Int64 != nil {
-			s = s + fmt.Sprintf(`dto.%[1]s = types.Int64Value(data["%[2]s"].(bool))`, CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
-			m = m + fmt.Sprintf("%[1]s         types.Int64 `json:\"%[2]s\"`", CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			s = s + fmt.Sprintf(`dto.%[1]s = types.Int64Value(data["%[2]s"].(bool))`, util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
+			m = m + fmt.Sprintf("%[1]s         types.Int64 `json:\"%[2]s\"`", util.ToPascalCase(n), PascalToSnakeCase(n)) + "\n"
 		} else if val.List != nil {
 			if val.List.ElementType.String != nil {
 				s = s + fmt.Sprintf(`"%[1]s": types.ListType{ElemType: types.StringType},`, n) + "\n"
@@ -41,10 +41,15 @@ func Gen_ConvertOAStoTFTypes(data resource.Attributes) (string, string, error) {
 			m = m + fmt.Sprintf("%[1]s         types.List `json:\"%[2]s\"`", CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
 		} else if val.SingleNested != nil {
 			s = s + fmt.Sprintf(`
-			temp%[1]s := data["%[2]s"].([]interface{})
+			temp%[1]s := data["%[2]s"].(map[string]interface{})
+			convertedTemp%[1]s, err := util.ConvertMapToObject(context.TODO(), temp%[1]s)
+			if err != nil {
+				log.Fatalf("ConvertMapToObject err: %v", err)
+			}
+
 			dto.%[1]s = diagOff(types.ObjectValueFrom, context.TODO(), types.ObjectType{AttrTypes: map[string]attr.Type{
 				%[3]s
-			}}}.AttributeTypes(), temp%[1]s)`, CamelToPascalCase(n), PascalToSnakeCase(n), GenObject(val.SingleNested.Attributes, n)) + "\n"
+			}}}.AttributeTypes(), convertedTemp%[1]s)`, CamelToPascalCase(n), PascalToSnakeCase(n), GenObject(val.SingleNested.Attributes, n)) + "\n"
 			m = m + fmt.Sprintf("%[1]s         types.Object `json:\"%[2]s\"`", CamelToPascalCase(n), PascalToSnakeCase(n)) + "\n"
 		}
 
@@ -82,16 +87,16 @@ func GenArray(d resource.Attributes, pName string) string {
 		n := val.Name
 
 		if val.String != nil {
-			t = t + fmt.Sprintf(`"%[1]s": types.StringType,`, n) + "\n"
+			t = t + fmt.Sprintf(`"%[1]s": types.StringType,`, util.ToCamelCase(n)) + "\n"
 		} else if val.Bool != nil {
-			t = t + fmt.Sprintf(`"%[1]s": types.BoolType,`, n) + "\n"
+			t = t + fmt.Sprintf(`"%[1]s": types.BoolType,`, util.ToCamelCase(n)) + "\n"
 		} else if val.Int64 != nil {
-			t = t + fmt.Sprintf(`"%[1]s": types.Int64Type,`, n) + "\n"
+			t = t + fmt.Sprintf(`"%[1]s": types.Int64Type,`, util.ToCamelCase(n)) + "\n"
 		} else if val.SingleNested != nil {
 			s = s + fmt.Sprintf(`
 			"%[1]s": types.ObjectType{AttrTypes: map[string]attr.Type{
 				%[2]s
-			}},`, n, GenObject(val.SingleNested.Attributes, n)) + "\n"
+			}},`, util.ToCamelCase(n), GenObject(val.SingleNested.Attributes, n)) + "\n"
 		}
 	}
 
@@ -111,27 +116,27 @@ func GenObject(d resource.Attributes, pName string) string {
 		n := val.Name
 
 		if val.String != nil {
-			s = s + fmt.Sprintf(`"%[1]s": types.StringType,`, n) + "\n"
+			s = s + fmt.Sprintf(`"%[1]s": types.StringType,`, util.ToCamelCase(n)) + "\n"
 		} else if val.Bool != nil {
-			s = s + fmt.Sprintf(`"%[1]s": types.BoolType,`, n) + "\n"
+			s = s + fmt.Sprintf(`"%[1]s": types.BoolType,`, util.ToCamelCase(n)) + "\n"
 		} else if val.Int64 != nil {
-			s = s + fmt.Sprintf(`"%[1]s": types.Int64Type,`, n) + "\n"
+			s = s + fmt.Sprintf(`"%[1]s": types.Int64Type,`, util.ToCamelCase(n)) + "\n"
 		} else if val.List != nil {
 			if val.List.ElementType.String != nil {
-				s = s + fmt.Sprintf(`"%[1]s": types.ListType{ElemType: types.StringType},`, n) + "\n"
+				s = s + fmt.Sprintf(`"%[1]s": types.ListType{ElemType: types.StringType},`, util.ToCamelCase(n)) + "\n"
 			} else if val.List.ElementType.Bool != nil {
-				s = s + fmt.Sprintf(`"%[1]s": types.ListType{ElemType: types.BoolType},`, n) + "\n"
+				s = s + fmt.Sprintf(`"%[1]s": types.ListType{ElemType: types.BoolType},`, util.ToCamelCase(n)) + "\n"
 			}
 		} else if val.ListNested != nil {
 			s = s + fmt.Sprintf(`
 			"%[1]s": types.ListType{ElemType:
 				%[2]s
-			},`, n, GenArray(val.ListNested.NestedObject.Attributes, n)) + "\n"
+			},`, util.ToCamelCase(n), GenArray(val.ListNested.NestedObject.Attributes, n)) + "\n"
 		} else if val.SingleNested != nil {
 			s = s + fmt.Sprintf(`
 			"%[1]s": types.ObjectType{AttrTypes: map[string]attr.Type{
 				%[2]s
-			}},`, n, GenObject(val.SingleNested.Attributes, n)) + "\n"
+			}},`, util.ToCamelCase(n), GenObject(val.SingleNested.Attributes, n)) + "\n"
 		}
 	}
 	return s
