+++ internal/util/codegen_helper.go
@@ -1,0 +1,101 @@
+package util
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/hashicorp/terraform-plugin-framework/attr"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+// Convert nested map structured json into terraform object
+func ConvertMapToObject(ctx context.Context, data map[string]interface{}) (types.Object, error) {
+	attrTypes := make(map[string]attr.Type)
+	attrValues := make(map[string]attr.Value)
+
+	for key, value := range data {
+		attrType, attrValue, err := convertInterfaceToAttr(ctx, value)
+		if err != nil {
+			return types.Object{}, fmt.Errorf("error converting field %s: %v", key, err)
+		}
+
+		attrTypes[key] = attrType
+		attrValues[key] = attrValue
+	}
+
+	r, _ := types.ObjectValue(attrTypes, attrValues)
+
+	return r, nil
+}
+
+// Convert interface{} into attr.Type attr.Value
+func convertInterfaceToAttr(ctx context.Context, value interface{}) (attr.Type, attr.Value, error) {
+	switch v := value.(type) {
+	case string:
+		return types.StringType, types.StringValue(v), nil
+
+	case float64:
+		return types.Int64Type, types.Int64Value(int64(v)), nil
+
+	case bool:
+		return types.BoolType, types.BoolValue(v), nil
+
+	case []interface{}:
+		if len(v) == 0 {
+			// Treat as array list in case of empty
+			return types.ListType{ElemType: types.StringType},
+				types.ListValueMust(types.StringType, []attr.Value{}),
+				nil
+		}
+
+		// Determine type based on first element
+		elemType, _, err := convertInterfaceToAttr(ctx, v[0])
+		if err != nil {
+			return nil, nil, err
+		}
+
+		values := make([]attr.Value, len(v))
+		for i, item := range v {
+			_, value, err := convertInterfaceToAttr(ctx, item)
+			if err != nil {
+				return nil, nil, err
+			}
+			values[i] = value
+		}
+
+		listType := types.ListType{ElemType: elemType}
+		listValue, diags := types.ListValue(elemType, values)
+		if diags.HasError() {
+			return nil, nil, err
+		}
+
+		return listType, listValue, nil
+
+	case map[string]interface{}:
+		objValue, err := ConvertMapToObject(ctx, v)
+		if err != nil {
+			return nil, nil, err
+		}
+		return objValue.Type(ctx), objValue, nil
+
+	case nil:
+		return types.StringType, types.StringNull(), nil
+
+	default:
+		return nil, nil, fmt.Errorf("unsupported type: %T", value)
+	}
+}
+
+func DiagOff[V, T interface{}](input func(ctx context.Context, elementType T, elements any) (V, diag.Diagnostics), ctx context.Context, elementType T, elements any) V {
+	var emptyReturn V
+
+	v, diags := input(ctx, elementType, elements)
+	fmt.Println(diags)
+	if diags.HasError() {
+		diags.AddError("REFRESING ERROR", "invalid diagOff operation")
+		return emptyReturn
+	}
+
+	return v
+}
